Move DefaultDeduced*
