+++ elevated-potential.core.impl/src/main/java/tripleo/elijah/stages/gen_c/internal/DefaultDeducedEvaConstructor.java
@@ -1,12 +1,12 @@
-package tripleo.elijah.stages.gen_c;
+package tripleo.elijah.stages.gen_c.internal;
 
 import io.reactivex.rxjava3.subjects.*;
-
 import org.jetbrains.annotations.*;
 import tripleo.elijah.*;
 import tripleo.elijah.lang.i.*;
 import tripleo.elijah.stages.deduce.*;
 import tripleo.elijah.stages.deduce.nextgen.*;
+import tripleo.elijah.stages.gen_c.*;
 import tripleo.elijah.stages.gen_fn.*;
 import tripleo.elijah.stages.gen_generic.*;
 import tripleo.elijah.stages.instructions.*;
